Remove previous versionof RecipeIngredients component, replace with Bootstrap version
